pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cipette"
version = "0.1.0"
description = "CI/CD Insights Dashboard - Collect and visualize GitHub Actions workflow metrics"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "novr", email = "novr@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "certifi>=2025.8.3",
    "flask>=3.0.0",
    "PyGithub>=2.1.1",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.1.0",
]
test = [
    "pytest>=8.0.0",
]

[project.scripts]
cipette-collect = "cipette.collector:main"
cipette-web = "cipette.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cipette*"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "C4", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests"
]

[dependency-groups]
dev = [
    "ruff>=0.13.2",
]
